Reworks process for generating and exporting CDS models
Previously each release of this module has exposed
only a single version of the CDS. However, at times
a consumer may want to use multiple versions of the
CDS concurrently e.g. the current version of the
Banking APIs but an older version of the Energy APIs.
This release changes the structure of the module
and is a breaking change. The package paths have
been changed. Each version of the CDS will be
under the url `github.com/fiskil/cdr/<industry>/<version>`.
This allows consumers to use multiple CDS versions
within the same module. Upgraders will need to update
their import paths accordingly.

The currently available CDS versions are 1.34.1 and
1.32.0